moving prior predictive to own function
